Updated minitoc's to only include sections.
Next chapter minitoc in particular was stretched over to the third page, messing up the formatting. However couldn't figure it out using `minitoc`, so changed to `etoc` which works this way out of the box.

But this change required some reformatting to couple chapters, mainly Z80. I must say the ordering of sections is better now! It also affected Next chapter which now has an empty page to the right of the minitoc. I contemplated moving the text between chapter title and minitoc there, to at least partially fill it up. But it's just couple lines so I decided there's no point... Maybe another idea will arrive in the future...