Updated `ADD rr,A` instruction effects table.
CF is undefined as discovered lately https://discord.com/channels/556228195767156758/695180116040351795/888099852725133412

Note: also realized `ADD` instruction was missing effects entries for Next extended variants. However this change pushed `AND` below it over to the next line, breaking the whole details chapter flow. I decided to rearrange `AND` content - bit table is now shown on the right of the description so the instruction uses less vertical space. Also changed `OR` and `XOR` in the same way. Spacing and page flow was fine there, but this way all instructions are similar.